# [Nim Con] improve performance