Here's a PR with ***only*** the `[Nim Con]` commits, i.e. affecting `nim_con` only, in case there's a hangup about the changes in `nim/` on `main` vs. `nim_rev` branches.

Will post fresh numbers from my Azure VM below. For `nim_con`, `nim_rev` consistently outperforms `main`, especially big differences for 20k and 60k runs.