@michaelsbradleyjr 2023-10-18
*WITHOUT* PGO
**Multi Threaded**
`nim_con` on `nim_rev` branch
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim_con.sh 
  main
* nim_con_rev
...
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 19.249ms
	Processing time (w/o IO): 19.071ms
	Processing time (w/o IO): 19.332ms
	Processing time (w/o IO): 19.68ms
	Processing time (w/o IO): 19.196ms
	Processing time (w/o IO): 19.294ms
	Processing time (w/o IO): 19.353ms
	Processing time (w/o IO): 19.337ms
	Processing time (w/o IO): 20.172ms
	Processing time (w/o IO): 19.256ms
	Processing time (w/o IO): 19.326ms
	Processing time (w/o IO): 19.163ms
	Processing time (w/o IO): 19.041ms
	 Time (mean ± σ):      43.2 ms ±   0.9 ms    [User: 64.9 ms, System: 8.9 ms]
	 Range (min … max):    41.2 ms …  43.9 ms    10 runs
	
Checking output
Success: related_posts_nim_con.json is valid!
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 253.505ms
	Processing time (w/o IO): 253.749ms
	Processing time (w/o IO): 253.7ms
	 Time (mean ± σ):     341.8 ms ±   0.4 ms    [User: 759.8 ms, System: 34.2 ms]
	 Range (min … max):   341.6 ms … 342.1 ms    2 runs
	
Running Nim Concurrent
using 4 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 2448.683ms
	Processing time (w/o IO): 2508.647ms
	Processing time (w/o IO): 2447.064ms
	 Time (mean ± σ):      2.760 s ±  0.039 s    [User: 6.968 s, System: 0.078 s]
	 Range (min … max):    2.732 s …  2.788 s    2 runs
```
*WITH* PGO
**Multi Threaded**
`nim_con` on `main` branch	 
```
(related) michael@azure1:~/repos/related_post_gen$ ./nim_con.sh 
* main
  nim_con_rev
...
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 31.299ms
	Processing time (w/o IO): 31.849ms
	Processing time (w/o IO): 33.879ms
	Processing time (w/o IO): 30.979ms
	Processing time (w/o IO): 31.164ms
	Processing time (w/o IO): 33.213ms
	Processing time (w/o IO): 31.885ms
	Processing time (w/o IO): 31.08ms
	Processing time (w/o IO): 31.684ms
	Processing time (w/o IO): 34.735ms
	Processing time (w/o IO): 33.313ms
	Processing time (w/o IO): 33.399ms
	Processing time (w/o IO): 31.201ms
	 Time (mean ± σ):      56.1 ms ±   1.5 ms    [User: 127.6 ms, System: 8.8 ms]
	 Range (min … max):    54.4 ms …  58.5 ms    10 runs
	
Checking output
Success: related_posts_nim_con.json is valid!
Running Nim Concurrent
using 8 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 448.425ms
	Processing time (w/o IO): 463.625ms
	Processing time (w/o IO): 475.878ms
	 Time (mean ± σ):     557.6 ms ±   8.6 ms    [User: 1716.6 ms, System: 36.1 ms]
	 Range (min … max):   551.5 ms … 563.6 ms    2 runs
	
Running Nim Concurrent
using 8 threads
Nim Concurrent:

	Benchmark 1: ./build/related_con
	Processing time (w/o IO): 4483.917ms
	Processing time (w/o IO): 4447.62ms
	Processing time (w/o IO): 4387.741ms
	 Time (mean ± σ):      4.694 s ±  0.046 s    [User: 15.380 s, System: 0.074 s]
	 Range (min … max):    4.662 s …  4.726 s    2 runs
```