@michaelsbradleyjr 2023-10-18
@inv2004 anything sticks out to you here as needing improvement? Obviously this project's maintainer has the say on what gets merged, but I always appreciate code review/suggestions from another Nim dev. Also, I'm experimenting with [taskpools](https://github.com/status-im/nim-taskpools) and [weave](https://github.com/mratsim/weave) as alternatives to [malebolgia](https://github.com/Araq/malebolgia) for use in `nim_con`. Any other ideas? You may be working on a totally different approach, and that would be cool too.