@jinyus 2023-10-18
```
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 21.642ms
        Processing time (w/o IO): 20.97ms
        Processing time (w/o IO): 20.944ms
        Processing time (w/o IO): 22.766ms
        Processing time (w/o IO): 21.653ms
        Processing time (w/o IO): 21.319ms
        Processing time (w/o IO): 21.419ms
        Processing time (w/o IO): 21.315ms
        Processing time (w/o IO): 21.183ms
        Processing time (w/o IO): 21.005ms
        Processing time (w/o IO): 21.795ms
        Processing time (w/o IO): 21.956ms
        Processing time (w/o IO): 21.708ms
          Time (mean ± σ):      49.9 ms ±   0.9 ms    [User: 73.1 ms, System: 10.9 ms]
          Range (min … max):    48.2 ms …  50.9 ms    10 runs
         
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 297.229ms
        Processing time (w/o IO): 294.407ms
        Processing time (w/o IO): 294.895ms
          Time (mean ± σ):     406.7 ms ±   1.1 ms    [User: 856.7 ms, System: 43.4 ms]
          Range (min … max):   405.9 ms … 407.5 ms    2 runs
         
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 2688.544ms
        Processing time (w/o IO): 2679.187ms
        Processing time (w/o IO): 2683.918ms
          Time (mean ± σ):      3.056 s ±  0.002 s    [User: 7.451 s, System: 0.113 s]
          Range (min … max):    3.055 s …  3.058 s    2 runs

```