@michaelsbradleyjr 2023-10-19
Sure, go for it.

I'm trying out `gcc` PGO vs. `clang` PGO, forgot to do that earlier per @inv2004's suggestion. If it makes a *consistently significant* difference it will be worth adding PGO steps back in, but that can be done in another PR.