@inv2004 2023-10-19
@michaelsbradleyjr Sorry for late response.

At the moment I think that it is hard to make any changes at all, because the VM results does not correlate to local results at all. The reasons for it could be in gcc/clang versions. Because I found that on old ubuntu clang+PGO is just terrible. + gcc is bad too